Added support for urlPrefix
Many CMSs (including Strapi) serve up non-absolute URLS
to the API (in the format '/image-url.jpg'). Currently the plugin
treats these as local images.

With the addition of a 'urlPrefix' config option, a full domain can be
provided which will be added to the target source URL before attempting
to download.